Simplify image file naming, rename service
