+++ api/images/models.py
@@ -1,7 +1,6 @@
 from django.db import models
 
 from api.common.mixins import ReprMixin
-from .utils import generate_image_path_partial
 
 
 class ImageAsset(ReprMixin, models.Model):
@@ -14,9 +13,8 @@ class ImageAsset(ReprMixin, models.Model):
         blank=True,
     )
     file = models.ImageField(
-        upload_to=generate_image_path_partial,
-        width_field="width",
-        height_field="height",
+        upload_to="images/", width_field="width", height_field="height",
     )
+
     height = models.PositiveIntegerField(blank=True, null=True)
     width = models.PositiveIntegerField(blank=True, null=True)
