+++ api/images/urls.py
@@ -4,6 +4,4 @@ from api.images import views
 urlpatterns = [
     # /images/...
     path("upload/", views.ImageAssetUploadView.as_view()),
-    # TODO Deprecate - use upload/
-    path("upload/<str:filename>", views.ImageAssetUploadView.as_view()),
 ]
